pyproject.toml
[project]
name = "structure_clip_app"
version = "0.1.0"
description = "Streamlit app using Structure-CLIP for multimodal structured representations"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "streamlit>=1.31",
  "torch>=2.4,<3",
  "torchvision>=0.19,<1.0",
  "transformers>=4.41",
  "accelerate>=0.30",
  "numpy>=1.23",
  "Pillow>=10.0",
  "tqdm>=4.66",
  "requests>=2.31",
  "protobuf>=4.25",
  "opencv-python-headless>=4.9",
  "sentencepiece>=0.1.99",
  "timm>=1.0.9",
  "setuptools>=68",
  "ftfy>=6.1.1",
  "regex>=2023.3.23",
]

[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/structure_clip_app"]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"


